Continue refactor

+++ index.js
@@ -2,23 +2,16 @@ const fs = require('fs')
 const moment = require('moment')
 const express = require('express')
 const open = require('open')
-const { COUNTER } = require('./vars')
+const { COUNTER, TOKEN_STORE } = require('./vars')
 
 const app = express()
 const port = 3000
 
-// create log directory (if DNE), today's log, and stats file (if DNE)
-const logDir = './twitch-logs'
+// create stats file & token store directory (if they don't exist)
+!fs.existsSync(`./${COUNTER}.json`) && fs.writeFileSync(`./${COUNTER}.json`, '{}')
+!fs.existsSync(`./${TOKEN_STORE}`) && fs.mkdirSync(`./${TOKEN_STORE}`)
+
 const startTime = moment()
-const dateString = startTime.format('YYYY-MM-DD_HH-mm-ss')
-const dateFilename = `${logDir}/${dateString}.txt`
-if (!fs.existsSync(logDir)) {
-  fs.mkdirSync(logDir)
-}
-fs.writeFileSync(dateFilename, `${dateString}\n\n`)
-if (!fs.existsSync(`./${COUNTER}.json`)) {
-  fs.writeFileSync(`./${COUNTER}.json`, '{}')
-}
 
 const initSpotify = require('./endpoints/init-spotify')
 const initTwitch = require('./endpoints/init-twitch')
@@ -30,7 +23,7 @@ app.get('/spotify.json', require('./endpoints/json-spotify'))
 app.get('/twitch.json', require('./endpoints/json-twitch'))
 app.get('/internal-stats.json', require('./endpoints/json-internal-stats'))
 app.get('/nintendo.json', require('./endpoints/json-nintendo'))
-app.get('/chat', require('./endpoints/chat')({ dateFilename, startTime }))
+app.get('/chat', require('./endpoints/chat')({ startTime }))
 
 app.listen(port, () => console.log(`Spotify callback API endpoint app listening on port ${port}.`))
-open('http://localhost:3000/chat')
+open(`${'http://localhost:3000'}${fs.existsSync(`./${TOKEN_STORE}/twitch-access`) ? '/chat' : '/inittoken-twitch'}`)
